@@ -1,95 +1,59 @@
 <template>
-  <div
-    v-show="menuShow"
-    class="contextmenu"
-    :style="{ top: menuTop + 'px', left: menuLeft + 'px' }"
-  >
+  <div v-show="show" class="contextmenu" :style="{ top: top + 'px', left: left + 'px' }">
     <ul>
-      <li v-if="curDeletable" @click="delWidgetObj">删除</li>
-      <li @click="topComponent">置顶</li>
-      <li @click="bottomComponent">置底</li>
-      <li @click="upComponent">上移</li>
-      <li @click="downComponent">下移</li>
+      <li @click="delWidgetObj">删除</li>
+      <li @click="setWidgetLock">锁定</li>
     </ul>
   </div>
 </template>
 
-<script>
-import { mapGetters } from 'vuex';
+<script setup lang="ts">
+  import { computed } from 'vue'
+  import { useStore } from '@/store'
+  const store = useStore()
 
-export default {
-  computed: {
-    ...mapGetters([
-      'menuTop',
-      'menuLeft',
-      'menuShow',
-      'rootWidgetObj',
-      'curDeletable',
-    ]),
-    isRootWidget() {
-      return false;
-    },
-  },
-  methods: {
-    delWidgetObj() {
-      this.$store.commit('delWidgetObj');
-      this.$store.commit('recordSnapshot');
-    },
+  const top = computed(() => store.getters['menu/top'])
+  const left = computed(() => store.getters['menu/left'])
+  const show = computed(() => store.getters['menu/show'])
 
-    upComponent() {
-      this.$store.commit('upComponent');
-      this.$store.commit('recordSnapshot');
-    },
+  const delWidgetObj = () => {
+    store.dispatch('widget/delete')
+    store.dispatch('menu/hidden')
+  }
 
-    downComponent() {
-      this.$store.commit('downComponent');
-      this.$store.commit('recordSnapshot');
-    },
-
-    topComponent() {
-      this.$store.commit('topComponent');
-      this.$store.commit('recordSnapshot');
-    },
-
-    bottomComponent() {
-      this.$store.commit('bottomComponent');
-      this.$store.commit('recordSnapshot');
-    },
-  },
-};
+  const setWidgetLock = () => {
+    store.dispatch('widget/lock')
+    store.dispatch('menu/hidden')
+  }
 </script>
 
 <style lang="scss" scoped>
-.contextmenu {
-  z-index: 1000;
-  position: fixed;
-
-  ul {
-    border: 1px solid #e4e7ed;
-    border-radius: 4px;
-    background-color: #fff;
-    box-shadow: 0 2px 12px 0 rgba(0, 0, 0, 0.1);
-    box-sizing: border-box;
-    margin: 5px 0;
-    padding: 6px 0;
-
-    li {
-      font-size: 14px;
-      padding: 0 20px;
-      position: relative;
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      color: #606266;
-      height: 34px;
-      line-height: 34px;
+  .contextmenu {
+    z-index: 1000;
+    position: fixed;
+    ul {
+      border: 1px solid #e4e7ed;
+      border-radius: 4px;
+      background-color: #fff;
+      box-shadow: 0 2px 12px 0 rgba(0, 0, 0, 0.1);
       box-sizing: border-box;
-      cursor: pointer;
-
-      &:hover {
-        background-color: #f5f7fa;
+      margin: 5px 0;
+      padding: 5px 0;
+      li {
+        font-size: 14px;
+        padding: 0 20px;
+        position: relative;
+        white-space: nowrap;
+        overflow: hidden;
+        text-overflow: ellipsis;
+        color: #606266;
+        height: 34px;
+        line-height: 34px;
+        cursor: pointer;
+        &:hover {
+          background-color: #f5f7fa;
+        }
       }
     }
   }
-}
 </style>
