@@ -1,18 +1,10 @@
 <template>
-  <div class="chart-list">
-    敬请期待！
-  </div>
+  <div class="chart-list"> 图表能力敬请期待！ </div>
 </template>
 
-<script>
-export default {
-  name: 'ChartList',
-};
-</script>
-
 <style lang="scss" scoped>
-.chart-list {
-  display: flex;
-  justify-content: center;
-}
+  .chart-list {
+    display: flex;
+    justify-content: center;
+  }
 </style>
