@@ -1,115 +1,115 @@
-import { getTag, isObject } from './type';
+import { getTag, isObject } from './type'
 
-const mapTag = '[object Map]';
-const setTag = '[object Set]';
-const arrayTag = '[object Array]';
-const objectTag = '[object Object]';
-const argsTag = '[object Arguments]';
-const boolTag = '[object Boolean]';
-const dateTag = '[object Date]';
-const numberTag = '[object Number]';
-const stringTag = '[object String]';
-const symbolTag = '[object Symbol]';
-const errorTag = '[object Error]';
-const regexpTag = '[object RegExp]';
-const funcTag = '[object Function]';
+const mapTag = '[object Map]'
+const setTag = '[object Set]'
+const arrayTag = '[object Array]'
+const objectTag = '[object Object]'
+const argsTag = '[object Arguments]'
+const boolTag = '[object Boolean]'
+const dateTag = '[object Date]'
+const numberTag = '[object Number]'
+const stringTag = '[object String]'
+const symbolTag = '[object Symbol]'
+const errorTag = '[object Error]'
+const regexpTag = '[object RegExp]'
+const funcTag = '[object Function]'
 
-const deepTag = [mapTag, setTag, arrayTag, objectTag, argsTag];
+const deepTag = [mapTag, setTag, arrayTag, objectTag, argsTag]
 
 function forEach(array, iteratee) {
-  let index = -1;
-  const length = array.length;
+  let index = -1
+  const length = array.length
   while (++index < length) {
-    iteratee(array[index], index);
+    iteratee(array[index], index)
   }
-  return array;
+  return array
 }
 
 function getInit(target) {
-  const Ctor = target.constructor;
-  return new Ctor();
+  const Ctor = target.constructor
+  return new Ctor()
 }
 
 function cloneSymbol(targe) {
-  return Object(Symbol.prototype.valueOf.call(targe));
+  return Object(Symbol.prototype.valueOf.call(targe))
 }
 
 function cloneReg(targe) {
-  const reFlags = /\w*$/;
-  const result = new targe.constructor(targe.source, reFlags.exec(targe));
-  result.lastIndex = targe.lastIndex;
-  return result;
+  const reFlags = /\w*$/
+  const result = new targe.constructor(targe.source, reFlags.exec(targe))
+  result.lastIndex = targe.lastIndex
+  return result
 }
 
 function cloneFunction(func) {
-  return func;
+  return func
 }
 
 function cloneOtherType(targe, type) {
-  const Ctor = targe.constructor;
+  const Ctor = targe.constructor
   switch (type) {
     case boolTag:
     case dateTag:
     case errorTag:
     case numberTag:
     case stringTag:
-      return new Ctor(targe);
+      return new Ctor(targe)
     case regexpTag:
-      return cloneReg(targe);
+      return cloneReg(targe)
     case symbolTag:
-      return cloneSymbol(targe);
+      return cloneSymbol(targe)
     case funcTag:
-      return cloneFunction(targe);
+      return cloneFunction(targe)
     default:
-      return null;
+      return null
   }
 }
 
 export function clone(target, map = new WeakMap()) {
   // 克隆原始类型
   if (!isObject(target)) {
-    return target;
+    return target
   }
 
   // 初始化
-  const type = getTag(target);
-  let cloneTarget;
+  const type = getTag(target)
+  let cloneTarget
   if (deepTag.includes(type)) {
-    cloneTarget = getInit(target, type);
+    cloneTarget = getInit(target, type)
   } else {
-    return cloneOtherType(target, type);
+    return cloneOtherType(target, type)
   }
 
   // 防止循环引用
   if (map.get(target)) {
-    return map.get(target);
+    return map.get(target)
   }
-  map.set(target, cloneTarget);
+  map.set(target, cloneTarget)
 
   // 克隆set
   if (type === setTag) {
     target.forEach((value) => {
-      cloneTarget.add(clone(value, map));
-    });
-    return cloneTarget;
+      cloneTarget.add(clone(value, map))
+    })
+    return cloneTarget
   }
 
   // 克隆map
   if (type === mapTag) {
     target.forEach((value, key) => {
-      cloneTarget.set(key, clone(value, map));
-    });
-    return cloneTarget;
+      cloneTarget.set(key, clone(value, map))
+    })
+    return cloneTarget
   }
 
   // 克隆对象和数组
-  const keys = type === arrayTag ? undefined : Object.keys(target);
+  const keys = type === arrayTag ? undefined : Object.keys(target)
   forEach(keys || target, (value, key) => {
     if (keys) {
-      key = value;
+      key = value
     }
-    cloneTarget[key] = clone(target[key], map);
-  });
+    cloneTarget[key] = clone(target[key], map)
+  })
 
-  return cloneTarget;
+  return cloneTarget
 }
