@@ -1,7 +1,5 @@
-
-export * from './type';
-export { clone } from './clone';
-export { eventBus } from './eventBus';
-export { generateID } from './generateID';
-export { debounce } from './debounce';
-export { throttle } from './throttle';
+export * from './type'
+export { clone } from './clone'
+export { generateID } from './generateID'
+export { debounce } from './debounce'
+export { throttle } from './throttle'
