@@ -1,5 +1,5 @@
-import { debounce } from './debounce';
-import { isObject } from './type';
+import { debounce } from './debounce'
+import { isObject } from './type'
 
 /**
  *
@@ -15,19 +15,19 @@ import { isObject } from './type';
  * @returns {Function} Returns the new throttled function.
  */
 export function throttle(func, wait, options) {
-  let leading = true;
-  let trailing = true;
+  let leading = true
+  let trailing = true
 
   if (typeof func !== 'function') {
-    throw new TypeError('Expected a function');
+    throw new TypeError('Expected a function')
   }
   if (isObject(options)) {
-    leading = 'leading' in options ? !!options.leading : leading;
-    trailing = 'trailing' in options ? !!options.trailing : trailing;
+    leading = 'leading' in options ? !!options.leading : leading
+    trailing = 'trailing' in options ? !!options.trailing : trailing
   }
   return debounce(func, wait, {
     leading,
     trailing,
-    'maxWait': wait,
-  });
+    maxWait: wait
+  })
 }
